+++ tests/unit/ScreenshotCommand.test.ts
@@ -1,0 +1,33 @@
+import "../helper/CfxGlobals";
+import "../helper/MockNatives";
+import "../helper/MockConfig";
+import { Logger } from "../../src/core/logger/Logger";
+import { Config } from "../../src/core/config/Config";
+import { ScreenshotCommand } from "../../src/commands/ScreenshotCommand";
+import { Utility } from "../../src/util/Utility";
+
+Logger.init();
+describe("ScreenshotCommand", () => {
+	Utility.EXPORTS["screenshot-basic"].requestClientScreenshot = jest.fn();
+	it("should handle invalid target argument", async () => {
+		global.GetResourceState = () => "started";
+		const command = new ScreenshotCommand();
+		await command.onExecute(1, ["invalid"]);
+		expect(Utility.EXPORTS["screenshot-basic"].requestClientScreenshot).not.toHaveBeenCalled();
+	});
+	it("should handle missing screenshot-basic resource", async () => {
+		global.GetResourceState = () => "stopped";
+		const command = new ScreenshotCommand();
+		await command.onExecute(1, ["1"]);
+		expect(Utility.EXPORTS["screenshot-basic"].requestClientScreenshot).not.toHaveBeenCalled();
+	});
+	it("should handle missing discord webhook", async () => {
+		global.GetResourceState = () => "started";
+		const command = new ScreenshotCommand();
+		const mockConfig = Config.getConfig();
+		mockConfig.DiscordWebhook = "";
+		jest.spyOn(Config, "getConfig").mockReturnValue(mockConfig);
+		await command.onExecute(1, ["1"]);
+		expect(Utility.EXPORTS["screenshot-basic"].requestClientScreenshot).not.toHaveBeenCalled();
+	});
+});
